requirements.txt
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
PyJWT==2.6.0
psycopg2-binary==2.9.5
